# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RoadSoS Kids Emergency Response System API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  activeIncidents: number;
  criticalIncidents: number;
  moderateIncidents: number;
  resolvedToday: number;
  averageResponseTimeMinutes: number;
  ambulancesDeployed: number;
  ambulancesAvailable: number;
  hospitalsAlerted: number;
  totalChildrenRegistered: number;
}